feat: Auth token interceptor on each request added
